Сделаны окна каталога во время заказа и заказа, и начата работа над страницей профиля

+++ PizzaMobileApp/pizzer_mobile/lib/blocs/bonuses_bloc/bonuses_events.dart
@@ -9,7 +9,7 @@ abstract class BonusesEvent extends Equatable {
 class BonusesChange extends BonusesEvent {
   final bool bonusesApplied;
   final int? userBonuses;
-  final double basketPrice;
+  final int basketPrice;
 
   BonusesChange(this.bonusesApplied, this.userBonuses, this.basketPrice);
   @override
